rename: features 폴더 명 소문자로 변경
